+++ sampler/src/main/java/atlantafx/sampler/page/components/ToggleSwitchPage.java
@@ -3,6 +3,7 @@
 package atlantafx.sampler.page.components;
 
 import atlantafx.base.controls.ToggleSwitch;
+import atlantafx.base.theme.Styles;
 import atlantafx.sampler.page.AbstractPage;
 import atlantafx.sampler.page.Page;
 import atlantafx.sampler.page.SampleBlock;
@@ -23,18 +24,19 @@ public class ToggleSwitchPage extends AbstractPage {
         super();
         setUserContent(new FlowPane(
             Page.PAGE_HGAP, Page.PAGE_VGAP,
-            basicSample()
+            basicSample(),
+            stateSample()
         ));
     }
 
     private SampleBlock basicSample() {
-        var leftToggle = new ToggleSwitch("Enable");
+        var leftToggle = new ToggleSwitch("Enabled");
         leftToggle.selectedProperty().addListener(
             (obs, old, val) -> leftToggle.setText(val ? "Enabled" : "Disabled")
         );
         leftToggle.setSelected(true);
 
-        var rightToggle = new ToggleSwitch("Disable");
+        var rightToggle = new ToggleSwitch("Disabled");
         rightToggle.selectedProperty().addListener(
             (obs, old, val) -> rightToggle.setText(val ? "Enabled" : "Disabled")
         );
@@ -42,5 +44,27 @@ public class ToggleSwitchPage extends AbstractPage {
         rightToggle.setSelected(false);
 
         return new SampleBlock("Basic", new VBox(SampleBlock.BLOCK_VGAP, leftToggle, rightToggle));
+    }
+
+    private SampleBlock stateSample() {
+        var successToggle = new ToggleSwitch("Enabled");
+        successToggle.selectedProperty().addListener((obs, old, val) -> {
+                successToggle.setText(val ? "Enabled" : "Disabled");
+                successToggle.pseudoClassStateChanged(Styles.STATE_SUCCESS, val);
+            }
+        );
+        successToggle.setSelected(true);
+        successToggle.pseudoClassStateChanged(Styles.STATE_SUCCESS, true);
+
+        var dangerToggle = new ToggleSwitch("Disabled");
+        dangerToggle.selectedProperty().addListener((obs, old, val) -> {
+                dangerToggle.setText(val ? "Enabled" : "Disabled");
+                dangerToggle.pseudoClassStateChanged(Styles.STATE_DANGER, val);
+            }
+        );
+        dangerToggle.setLabelPosition(HorizontalDirection.RIGHT);
+        dangerToggle.setSelected(false);
+
+        return new SampleBlock("State", new VBox(SampleBlock.BLOCK_VGAP, successToggle, dangerToggle));
     }
 }