+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - (Base) Improved Javadoc. See full API reference in [docs](https://mkpaz.github.io/atlantafx/apidocs/atlantafx.base/module-summary.html).
 - (Base) `ToggleSwitch` label position support (left or right).
+- (CSS) `ToggleSwitch` pseudo-classes (`success`, `danger`) support.
 - (CSS) `Button` shadow support (`-color-button-shadow`). Only for themes compiled with the `button.$use-shadow` flag enabled.  
 - (CSS) Looked-up color variables support: `Separator`.
 - (CSS) Added border radius/shadow to popup menu for `ComboBox` (and all `ComboBox`-based) controls.